r13802: I *knew* ASU on sparc had to be good for *something* ! :-). Fix incorrect size understanding of sid name type (yes it's already correct in the Samba4 IDL :-). Jeremy.

+++ source/include/rpc_misc.h
@@ -266,14 +266,14 @@ typedef struct {
 
 /* DOM_RID - domain RID structure for ntlsa pipe */
 typedef struct {
-	uint8 type; /* value is SID_NAME_USE enum */
+	uint16 type; /* value is SID_NAME_USE enum */
 	uint32 rid;
 	uint32 rid_idx; /* referenced domain index */
 } DOM_RID;
 
 /* DOM_RID2 - second domain RID structure for ntlsa pipe */
 typedef struct {
-	uint8 type; /* value is SID_NAME_USE enum */
+	uint16 type; /* value is SID_NAME_USE enum */
 	uint32 rid;
 	uint32 rid_idx; /* referenced domain index */
 	uint32 unknown;
